PHP MySQL Workbench Connection
First you need to create the database in MySQL Workbench. Then you can fill in your own sample data if you want. Next go the the "dbconnection.php" file and make any changes that are specific to your database (this would probably be the password for localhost and the database name). I also need to change the other files to have the php file name so they can work together. You also need a few extensions to get it to work, but I can help you with that when we talk together.