Added Shopping Cart Functionality
Items can now be added to the shopping cart. Uses cookies to store item id's of items wanted. then queries the other information to be displayed on the "Buy_Page.php" or "View Cart" pages. Also has a running sum for total price. No actual purchasing functionality yet. Also, I need to make it where once an item is in the uses cart, the "add to cart" button changes to show that that item is in the users cart.