Added Ability to Remove Items from Cart
Items on the shopping cart page now have a red "remove from cart" button. once pressed it removes the item and refreshes the page. Also, when adding an item to the cart on one of the other pages, the green "add to cart" button changes into a "remove from cart" button to show that the user has indeed added that item to the cart, as well as let them remove it from their cart without being on the shopping cart page.